integrated casl into frontend for permission based access control, protected routes and elements based on roles and permissions
