+++ app/App.tsx
@@ -1,8 +1,15 @@
+import { QueryClientProvider } from '@tanstack/react-query';
+import AuthProvider from './context/AuthContext';
 import Root from './routes/root';
+import { queryClient } from './utils/query-client';
 
 function App() {
   return (
-    <Root />
+    <QueryClientProvider client={queryClient}>
+      <AuthProvider>
+        <Root />
+      </AuthProvider>
+    </QueryClientProvider>
   );
 }
