+++ app/features/team/components/TeamsEmptyState.tsx
@@ -1,11 +1,20 @@
 import { Search, Plus } from 'lucide-react';
 import { Button } from '@app/components/ui/button';
+import { useAbility } from '@casl/react';
+import { AbilityContext, Can } from '@app/casl/AbilityContext';
 
 interface TeamsEmptyStateProps {
   onCreateTeam?: () => void;
+  isLoading: boolean;
 }
 
-export const TeamsEmptyState = ({ onCreateTeam }: TeamsEmptyStateProps) => {
+export const TeamsEmptyState = ({ onCreateTeam, isLoading }: TeamsEmptyStateProps) => {
+  const ability = useAbility(AbilityContext);
+
+  if (isLoading) {
+    return;
+  }
+
   return (
     <div className="flex flex-col items-center justify-center py-12">
       <div className="text-center">
@@ -15,12 +24,14 @@ export const TeamsEmptyState = ({ onCreateTeam }: TeamsEmptyStateProps) => {
           Try adjusting your search criteria or create a new team.
         </p>
 
-        {onCreateTeam && (
-          <Button onClick={onCreateTeam} className="mt-4">
-            <Plus className="mr-2 h-4 w-4" />
-            Create Team
-          </Button>
-        )}
+        <Can I="create" a="team" ability={ability}>
+          {onCreateTeam && (
+            <Button onClick={onCreateTeam} className="mt-4">
+              <Plus className="mr-2 h-4 w-4" />
+              Create Team
+            </Button>
+          )}
+        </Can>
       </div>
     </div>
   );
