+++ app/pages/Buckets.tsx
@@ -8,6 +8,8 @@ import { Button } from '@app/components/ui/button.tsx';
 import { useFilteredBuckets } from '@app/features/buckets/hooks/useFilteredBuckets.ts';
 import { useFilteredUserBuckets } from '@app/features/buckets/hooks/useFilteredUserBuckets.ts';
 import { AddBucketDialog } from '@app/features/buckets/components/AddBucketDialog.tsx';
+import { useAbility } from '@casl/react';
+import { AbilityContext, Can } from '@app/casl/AbilityContext.ts';
 
 function getMaxLevelForCategory(categories: BucketCategory[] | undefined, categoryId: string) {
   const category = categories && categories.find((cat) => cat.id === categoryId);
@@ -16,6 +18,8 @@ function getMaxLevelForCategory(categories: BucketCategory[] | undefined, catego
 }
 
 export const Buckets = () => {
+  const ability = useAbility(AbilityContext);
+
   const { userBuckets, isLoading } = useGetMyUserBuckets();
   const { categories } = useGetCategories();
 
@@ -66,8 +70,8 @@ export const Buckets = () => {
         </div>
 
         {/* Create Bucket Button and View Mode Toggle */}
-        <div className="ml-auto flex items-center justify-end gap-2">
-          <div className="mr-3 flex rounded-lg border-1">
+        <div className="ml-auto flex items-center justify-end gap-3">
+          <div className="flex rounded-lg border-1">
             <Button
               size="icon"
               onClick={handleSetListMode}
@@ -85,10 +89,12 @@ export const Buckets = () => {
             </Button>
           </div>
 
-          <Button onClick={handleOpenAddBucketModal} className="flex items-center gap-2">
-            <Plus className="h-4 w-4" />
-            Create Bucket
-          </Button>
+          <Can I="create" a="BucketCategory" ability={ability}>
+            <Button onClick={handleOpenAddBucketModal} className="flex items-center gap-2">
+              <Plus className="h-4 w-4" />
+              Create Bucket
+            </Button>
+          </Can>
         </div>
       </div>
