+++ app/pages/ReportDetail.tsx
@@ -1,10 +1,9 @@
-import { useParams, useNavigate } from 'react-router-dom';
+import { useParams, useNavigate, Navigate } from 'react-router-dom';
 import { Avatar, AvatarFallback, AvatarImage } from '../components/ui/avatar';
 import { Badge } from '../components/ui/badge';
 import { Button } from '../components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../components/ui/card';
 import { useGetReportById } from '@app/hooks/report';
-import { Report } from '@app/types/types';
 import { Mail, Calendar, User, ArrowLeft, FileText, Edit, Trash2 } from 'lucide-react';
 import { Spinner } from '@app/components/ui/spinner';
 import routeNames from '@app/routes/route-names';
@@ -13,16 +12,22 @@ import { useDeleteReport } from '@app/hooks/report';
 import { useGetUserById } from '@app/hooks/user';
 import { toast } from 'sonner';
 import { useAuth } from '@app/context/AuthContext';
+import { AbilityContext, Can } from '@app/casl/AbilityContext';
+import { useAbility } from '@casl/react';
 
 export const ReportDetail = () => {
   const { reportId } = useParams<{ reportId: string }>();
+  const ability = useAbility(AbilityContext);
+
+  if (ability.cannot('read', 'Reports')) {
+    <Navigate to={routeNames.notAuthorized()} />;
+  }
+
   const navigate = useNavigate();
+  const { user: currentUser } = useAuth();
   const { report, isLoading, isSuccess, error } = useGetReportById(reportId || '');
   const { mutate: deleteReport } = useDeleteReport();
   const [isDeleting, setIsDeleting] = useState(false);
-  const { user: currentUser } = useAuth();
-
-
 
   // Get user data for the report
   const { user: reportUser, isLoading: userLoading } = useGetUserById(report?.userId || '');
@@ -101,7 +106,9 @@ export const ReportDetail = () => {
       <div className="flex h-full items-center justify-center">
         <div className="text-center">
           <h2 className="mb-4 text-2xl font-bold">Report Not Found</h2>
-          <p className="mb-4 text-gray-600">The report you're looking for doesn't exist or you don't have permission to view it.</p>
+          <p className="mb-4 text-gray-600">
+            The report you're looking for doesn't exist or you don't have permission to view it.
+          </p>
           <Button onClick={handleBackButton}>
             <ArrowLeft className="mr-2 h-4 w-4" />
             Back to Reports
@@ -130,29 +137,30 @@ export const ReportDetail = () => {
             </div>
             {isAuthor && (
               <div className="flex items-center gap-2">
-                <Button
-                  variant="outline"
-                  size="sm"
-                  onClick={handleEditReport}
-                  className="flex items-center gap-2"
-                >
-                  <Edit className="h-4 w-4" />
-                  Edit
-                </Button>
-                <Button
-                  variant="destructive"
-                  size="sm"
-                  onClick={handleDeleteReport}
-                  disabled={isDeleting}
-                  className="flex items-center gap-2"
-                >
-                  {isDeleting ? (
-                    <Spinner size="small" />
-                  ) : (
-                    <Trash2 className="h-4 w-4" />
-                  )}
-                  Delete
-                </Button>
+                <Can I="update" a="Reports" ability={ability}>
+                  <Button
+                    variant="outline"
+                    size="sm"
+                    onClick={handleEditReport}
+                    className="flex items-center gap-2"
+                  >
+                    <Edit className="h-4 w-4" />
+                    Edit
+                  </Button>
+                </Can>
+
+                <Can I="delete" a="Reports" ability={ability}>
+                  <Button
+                    variant="destructive"
+                    size="sm"
+                    onClick={handleDeleteReport}
+                    disabled={isDeleting}
+                    className="flex items-center gap-2"
+                  >
+                    {isDeleting ? <Spinner size="small" /> : <Trash2 className="h-4 w-4" />}
+                    Delete
+                  </Button>
+                </Can>
               </div>
             )}
           </div>
@@ -169,9 +177,7 @@ export const ReportDetail = () => {
                 <FileText className="h-5 w-5" />
                 Report Content
               </CardTitle>
-              <CardDescription>
-                Detailed feedback and evaluation
-              </CardDescription>
+              <CardDescription>Detailed feedback and evaluation</CardDescription>
             </CardHeader>
             <CardContent>
               <div className="prose max-w-none">
@@ -195,14 +201,14 @@ export const ReportDetail = () => {
             </CardHeader>
             <CardContent className="space-y-4">
               <div className="flex items-center gap-2">
-                <Calendar className="h-4 w-4 text-muted-foreground" />
-                <span className="text-sm text-muted-foreground">Created</span>
+                <Calendar className="text-muted-foreground h-4 w-4" />
+                <span className="text-muted-foreground text-sm">Created</span>
                 <span className="text-sm font-medium">{formatDate(report.createdAt)}</span>
               </div>
               {report.updatedAt !== report.createdAt && (
                 <div className="flex items-center gap-2">
-                  <Calendar className="h-4 w-4 text-muted-foreground" />
-                  <span className="text-sm text-muted-foreground">Updated</span>
+                  <Calendar className="text-muted-foreground h-4 w-4" />
+                  <span className="text-muted-foreground text-sm">Updated</span>
                   <span className="text-sm font-medium">{formatDate(report.updatedAt)}</span>
                 </div>
               )}
@@ -249,54 +255,58 @@ export const ReportDetail = () => {
             </CardHeader>
             <CardContent>
               <div className="space-y-4">
-                                  <div className="flex items-center gap-3">
-                    <Avatar className="h-12 w-12">
-                      <AvatarImage src="" />
-                      <AvatarFallback className="text-primary-foreground bg-neutral-800 text-lg font-semibold">
-                        {reportUser ? getInitials(reportUser.firstName, reportUser.lastName) : 'U'}
-                      </AvatarFallback>
-                    </Avatar>
-                    <div>
-                      <h3 className="text-foreground text-lg font-semibold">
-                        {reportUser ? `${reportUser.firstName} ${reportUser.lastName}` : `User ${report.userId}`}
-                      </h3>
-                      {reportUser && (
-                        <div className="flex items-center gap-2 mt-1">
-                          <Badge
-                            variant="secondary"
-                            className={`${getStatusBadgeClass(reportUser.status)} text-xs`}
-                          >
-                            {reportUser.status}
-                          </Badge>
-                        </div>
-                      )}
-                    </div>
+                <div className="flex items-center gap-3">
+                  <Avatar className="h-12 w-12">
+                    <AvatarImage src="" />
+                    <AvatarFallback className="text-primary-foreground bg-neutral-800 text-lg font-semibold">
+                      {reportUser ? getInitials(reportUser.firstName, reportUser.lastName) : 'U'}
+                    </AvatarFallback>
+                  </Avatar>
+                  <div>
+                    <h3 className="text-foreground text-lg font-semibold">
+                      {reportUser
+                        ? `${reportUser.firstName} ${reportUser.lastName}`
+                        : `User ${report.userId}`}
+                    </h3>
+                    {reportUser && (
+                      <div className="mt-1 flex items-center gap-2">
+                        <Badge
+                          variant="secondary"
+                          className={`${getStatusBadgeClass(reportUser.status)} text-xs`}
+                        >
+                          {reportUser.status}
+                        </Badge>
+                      </div>
+                    )}
                   </div>
-                
+                </div>
+
                 {reportUser && (
                   <div className="space-y-2">
                     <div className="flex items-center gap-2">
-                      <Mail className="h-4 w-4 text-muted-foreground" />
-                      <span className="text-sm text-muted-foreground">Email</span>
+                      <Mail className="text-muted-foreground h-4 w-4" />
+                      <span className="text-muted-foreground text-sm">Email</span>
                       <span className="text-sm font-medium">{reportUser.email}</span>
                     </div>
                     {reportUser.phoneNumber && (
                       <div className="flex items-center gap-2">
-                        <Mail className="h-4 w-4 text-muted-foreground" />
-                        <span className="text-sm text-muted-foreground">Phone</span>
+                        <Mail className="text-muted-foreground h-4 w-4" />
+                        <span className="text-muted-foreground text-sm">Phone</span>
                         <span className="text-sm font-medium">{reportUser.phoneNumber}</span>
                       </div>
                     )}
                   </div>
                 )}
 
-                <Button
-                  onClick={() => handleViewUser(report.userId)}
-                  className="w-full"
-                  variant="outline"
-                >
-                  View User Profile
-                </Button>
+                <Can I="read" a="User" ability={ability}>
+                  <Button
+                    onClick={() => handleViewUser(report.userId)}
+                    className="w-full"
+                    variant="outline"
+                  >
+                    View User Profile
+                  </Button>
+                </Can>
               </div>
             </CardContent>
           </Card>
@@ -304,4 +314,4 @@ export const ReportDetail = () => {
       </div>
     </div>
   );
-}; 
+};
