+++ app/pages/TicketEmployee.tsx
@@ -1,41 +1,34 @@
-import { useState, useEffect } from 'react';
-import { getAllUsersApi } from '@app/api/user-api';
-import { useAuth } from '@app/context/AuthContext';
-import { User } from '@app/types/types';
-import { Ticket } from '@app/types/ticket';
-import { CreateTicketValue } from '@app/schemas/ticketSchema';
+import { useState } from 'react';
+import { useAuth } from '../context/AuthContext';
 import {
-  useRealtimeAllTickets,
   useRealtimeChat,
+  useRealtimeMyTickets,
   useRealtimeTicket,
 } from '@app/features/tickets/hooks';
-import { Button } from '@app/components/ui/button';
-import { Plus, X } from 'lucide-react';
+import { Ticket } from '@app/types/ticket';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@app/components/ui/select';
 
-// Interfaces now imported from @app/features/tickets
-
-export const CtoTicketTest = () => {
+export const TicketEmployee = () => {
   const { user: currentUser } = useAuth();
-  const [users, setUsers] = useState<User[]>([]);
-  const [employeeId, setEmployeeId] = useState<string>('');
   const [selectedTicket, setSelectedTicket] = useState<Ticket | null>(null);
-  const [showCreateForm, setShowCreateForm] = useState(false);
+  const [searchTerm, setSearchTerm] = useState('');
+  const [sortBy, setSortBy] = useState<'date' | 'title'>('date');
   const [expandedDescriptions, setExpandedDescriptions] = useState<Set<string>>(new Set());
-  const [createForm, setCreateForm] = useState<CreateTicketValue>({
-    title: '',
-    description: '',
-  });
 
   // Use real-time ticket management
-  const { tickets, ticketCounts, isLoading, createNewTicket, isCreating, hasAwaitingConfirmation } =
-    useRealtimeAllTickets();
+  const { tickets, isLoading } = useRealtimeMyTickets();
 
   // Use real-time selected ticket details
   const {
     ticket: realtimeTicket,
-    isConfirming,
-    markTicketFinished,
-    isMarkingFinished,
+    markAsAwaitingConfirmation,
+    isMarkingAsFinished,
     canMarkFinished,
   } = useRealtimeTicket(selectedTicket?.id || null);
 
@@ -54,54 +47,13 @@ export const CtoTicketTest = () => {
 
   const [newMessage, setNewMessage] = useState('');
 
-  useEffect(() => {
-    loadUsers();
-  }, []);
-
-  const loadUsers = async () => {
-    try {
-      const userData = await getAllUsersApi();
-      if (userData) {
-        setUsers(userData);
-      }
-    } catch (error) {
-      console.error('Failed to load users:', error);
-    }
-  };
-
-  const handleCreateTicket = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    console.log('Creating Ticket: ', { createForm });
-    try {
-      await createNewTicket(employeeId, createForm);
-      setCreateForm({ title: '', description: '' });
-      setShowCreateForm(false);
-      console.log('Ticket created successfully!');
-    } catch (error) {
-      console.error('Failed to create ticket:', error);
-    }
-  };
-
-  const handleSendMessage = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!newMessage.trim() || !currentTicket) return;
-
-    try {
-      await sendMessage(newMessage);
-      setNewMessage('');
-    } catch (error) {
-      console.error('Failed to send message:', error);
-    }
-  };
-
   const handleMarkAsFinished = async () => {
     if (!currentTicket || !canMarkFinished) return;
 
     try {
-      await markTicketFinished(currentTicket.id);
+      await markAsAwaitingConfirmation(currentTicket.id);
     } catch (error) {
-      console.error('Failed to mark ticket as finished by CTO:', error);
+      console.error('Failed to mark ticket as finished:', error);
     }
   };
 
@@ -113,25 +65,6 @@ export const CtoTicketTest = () => {
       newExpanded.add(ticketId);
     }
     setExpandedDescriptions(newExpanded);
-  };
-
-  const formatDateTime = (dateString: string) => {
-    return new Date(dateString).toLocaleString();
-  };
-
-  const formatTime = (dateString: string) => {
-    return new Date(dateString).toLocaleTimeString([], {
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  };
-
-  const getInitials = (name: string) => {
-    return name
-      .split(' ')
-      .map((n) => n[0])
-      .join('')
-      .toUpperCase();
   };
 
   const getStatusColor = (status: string) => {
@@ -179,245 +112,146 @@ export const CtoTicketTest = () => {
     }
   };
 
+  const handleSendMessage = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!newMessage.trim() || !selectedTicket) return;
+
+    try {
+      await sendMessage(newMessage);
+      setNewMessage('');
+    } catch (error) {
+      console.error('Failed to send message:', error);
+    }
+  };
+
+  const formatDateTime = (dateString: string) => {
+    return new Date(dateString).toLocaleString();
+  };
+
+  const formatTime = (dateString: string) => {
+    return new Date(dateString).toLocaleTimeString([], {
+      hour: '2-digit',
+      minute: '2-digit',
+    });
+  };
+
+  const getInitials = (name: string) => {
+    return name
+      .split(' ')
+      .map((n) => n[0])
+      .join('')
+      .toUpperCase();
+  };
+
   // Fix message ownership detection
   const isMyMessage = (senderId: string) => {
     if (!currentUser) return false;
     return senderId === currentUser.id;
   };
 
+  const getTimeAgo = (dateString: string) => {
+    const now = new Date();
+    const date = new Date(dateString);
+    const diffInMinutes = Math.floor((now.getTime() - date.getTime()) / (1000 * 60));
+
+    if (diffInMinutes < 60) {
+      return `${diffInMinutes} min ago`;
+    } else if (diffInMinutes < 1440) {
+      return `${Math.floor(diffInMinutes / 60)} hours ago`;
+    } else {
+      return `${Math.floor(diffInMinutes / 1440)} days ago`;
+    }
+  };
+
+  const filteredTickets = tickets
+    .filter(
+      (ticket) =>
+        ticket.title.toLowerCase().includes(searchTerm.toLowerCase()) ||
+        ticket.description.toLowerCase().includes(searchTerm.toLowerCase()),
+    )
+    .sort((a, b) => {
+      if (sortBy === 'date') {
+        return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
+      } else {
+        return a.title.localeCompare(b.title);
+      }
+    });
+
   if (isLoading) {
     return (
       <div className="flex min-h-screen items-center justify-center bg-gray-50">
         <div className="text-center">
           <div className="mx-auto h-12 w-12 animate-spin rounded-full border-b-2 border-blue-500"></div>
-          <p className="mt-4 text-gray-600">Loading tickets...</p>
+          <p className="mt-4 text-gray-600">Loading your tickets...</p>
         </div>
       </div>
     );
   }
 
   return (
-    <div className="h-full bg-gray-100">
+    <div className="min-h-screen bg-gray-100">
       <div className="mx-auto max-w-7xl p-6">
         <div className="mb-8">
-          <h1 className="mb-2 text-3xl font-bold text-gray-900">Ticket Management</h1>
-          <p className="text-gray-600">Create, assign, and manage team tickets</p>
+          <h1 className="mb-2 text-3xl font-bold text-gray-900">My Assigned Tickets</h1>
+          <p className="text-gray-600">Manage and track your assigned work tickets</p>
 
-          <div className="mt-6">
-            <Button
-              variant={showCreateForm ? 'outline' : 'default'}
-              size="lg"
-              onClick={() => setShowCreateForm((prev) => !prev)}
-              className={`${showCreateForm ? 'border-primary/30' : ''} w-46`}
-            >
-              {showCreateForm ? (
-                <>
-                  <X /> Cancel
-                </>
-              ) : (
-                <>
-                  <Plus /> Create New Ticket
-                </>
-              )}
-            </Button>
+          <div className="mt-6 flex gap-4">
+            <div className="flex-1">
+              <div className="relative">
+                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+                  <svg
+                    className="h-5 w-5 text-gray-400"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
+                    />
+                  </svg>
+                </div>
+                <input
+                  type="text"
+                  placeholder="Search tickets..."
+                  value={searchTerm}
+                  onChange={(e) => setSearchTerm(e.target.value)}
+                  className="border-primary/30 bg-card h-9 w-full rounded-lg border pr-4 pl-10 transition-colors focus:border-blue-500 focus:ring-2 focus:ring-blue-500"
+                />
+              </div>
+            </div>
+            <Select>
+              <SelectTrigger className="bg-card border-primary/30 w-34">
+                <SelectValue defaultValue={sortBy} placeholder="Sort by Date" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="date" onClick={() => setSortBy('date')}>
+                  Sort by Date
+                </SelectItem>
+                <SelectItem value="title" onClick={() => setSortBy('date')}>
+                  Sort by Title
+                </SelectItem>
+              </SelectContent>
+            </Select>
           </div>
         </div>
-
-        {showCreateForm && (
-          <div className="mb-8 rounded-xl border border-gray-200 bg-white shadow-sm">
-            <div className="border-b border-gray-100 p-6">
-              <h2 className="text-xl font-semibold text-gray-900">Create New Ticket</h2>
-              <p className="mt-1 text-sm text-gray-600">Assign a new task to a team member</p>
-            </div>
-
-            <form onSubmit={handleCreateTicket} className="p-6">
-              <div className="grid grid-cols-1 gap-6 sm:grid-cols-2">
-                <div className="sm:col-span-2">
-                  <label className="mb-2 block text-sm font-medium text-gray-700">
-                    Ticket Title
-                  </label>
-                  <input
-                    type="text"
-                    value={createForm.title}
-                    onChange={(e) => setCreateForm({ ...createForm, title: e.target.value })}
-                    className="w-full rounded-lg border border-gray-300 px-4 py-3 transition-colors focus:border-blue-500 focus:ring-2 focus:ring-blue-500"
-                    placeholder="Enter a clear, descriptive title"
-                    required
-                  />
-                </div>
-
-                <div className="sm:col-span-2">
-                  <label className="mb-2 block text-sm font-medium text-gray-700">
-                    Description
-                  </label>
-                  <textarea
-                    value={createForm.description}
-                    onChange={(e) => setCreateForm({ ...createForm, description: e.target.value })}
-                    rows={4}
-                    className="w-full rounded-lg border border-gray-300 px-4 py-3 transition-colors focus:border-blue-500 focus:ring-2 focus:ring-blue-500"
-                    placeholder="Provide detailed information about the task..."
-                    required
-                  />
-                </div>
-
-                <div className="sm:col-span-2">
-                  <label className="mb-2 block text-sm font-medium text-gray-700">
-                    Assign to Employee
-                  </label>
-                  <select
-                    value={employeeId}
-                    onChange={(e) => setEmployeeId(e.target.value)}
-                    className="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 transition-colors focus:border-blue-500 focus:ring-2 focus:ring-blue-500"
-                    required
-                  >
-                    <option value="">Select an employee</option>
-                    {users.map((user) => (
-                      <option key={user.id} value={user.id}>
-                        {user.firstName} {user.lastName} ({user.email})
-                      </option>
-                    ))}
-                  </select>
-                </div>
-              </div>
-
-              <div className="mt-6 flex items-center gap-3">
-                <button
-                  type="submit"
-                  disabled={isCreating}
-                  className="inline-flex items-center rounded-lg bg-green-600 px-6 py-3 font-medium text-white transition-colors hover:bg-green-700 focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50"
-                >
-                  {isCreating ? (
-                    <>
-                      <svg
-                        className="mr-2 -ml-1 h-5 w-5 animate-spin text-white"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                      >
-                        <circle
-                          className="opacity-25"
-                          cx="12"
-                          cy="12"
-                          r="10"
-                          stroke="currentColor"
-                          strokeWidth="4"
-                        ></circle>
-                        <path
-                          className="opacity-75"
-                          fill="currentColor"
-                          d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                        ></path>
-                      </svg>
-                      Creating Ticket...
-                    </>
-                  ) : (
-                    <>
-                      <svg
-                        className="mr-2 h-5 w-5"
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth="2"
-                          d="M12 6v6m0 0v6m0-6h6m-6 0H6"
-                        />
-                      </svg>
-                      Create Ticket
-                    </>
-                  )}
-                </button>
-                <button
-                  type="button"
-                  onClick={() => setShowCreateForm(false)}
-                  className="inline-flex items-center rounded-lg bg-gray-600 px-6 py-3 font-medium text-white transition-colors hover:bg-gray-700 focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 focus:outline-none"
-                >
-                  Cancel
-                </button>
-              </div>
-            </form>
-          </div>
-        )}
-
-        {/* Ticket Status Overview */}
-        {hasAwaitingConfirmation && (
-          <div className="mb-6 rounded-lg border border-amber-200 bg-amber-50 p-4">
-            <div className="flex items-center">
-              <div className="flex-shrink-0">
-                <svg
-                  className="h-5 w-5 text-amber-400"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z"
-                  />
-                </svg>
-              </div>
-              <div className="ml-3">
-                <h3 className="text-sm font-medium text-amber-800">
-                  {ticketCounts.awaitingConfirmation} ticket(s) need your confirmation
-                </h3>
-                <p className="text-sm text-amber-700">
-                  Employees have marked these tickets as finished and are waiting for your approval.
-                </p>
-              </div>
-            </div>
-          </div>
-        )}
 
         <div className="grid grid-cols-1 gap-8 lg:grid-cols-2">
           {/* Tickets List */}
           <div className="rounded-xl border border-gray-200 bg-white shadow-sm">
             <div className="border-b border-gray-100 p-6">
               <div className="flex items-center justify-between">
-                <h2 className="text-xl font-semibold text-gray-900">All Tickets</h2>
-                <div className="flex items-center gap-3">
-                  {hasAwaitingConfirmation && (
-                    <span className="inline-flex items-center rounded-full bg-amber-100 px-2.5 py-0.5 text-xs font-medium text-amber-800">
-                      🔔 {ticketCounts.awaitingConfirmation} need confirmation
-                    </span>
-                  )}
-                  <span className="inline-flex items-center rounded-full bg-blue-100 px-2.5 py-0.5 text-xs font-medium text-blue-800">
-                    {tickets.length} total
-                  </span>
-                </div>
-              </div>
-
-              {/* Status breakdown */}
-              <div className="mt-4 grid grid-cols-4 gap-4">
-                <div className="text-center">
-                  <div className="text-lg font-semibold text-yellow-600">
-                    {ticketCounts.pending}
-                  </div>
-                  <div className="text-xs text-gray-500">Pending</div>
-                </div>
-                <div className="text-center">
-                  <div className="text-lg font-semibold text-blue-600">{ticketCounts.ongoing}</div>
-                  <div className="text-xs text-gray-500">Ongoing</div>
-                </div>
-                <div className="text-center">
-                  <div className="text-lg font-semibold text-purple-600">
-                    {ticketCounts.awaitingConfirmation}
-                  </div>
-                  <div className="text-xs text-gray-500">Awaiting</div>
-                </div>
-                <div className="text-center">
-                  <div className="text-lg font-semibold text-green-600">
-                    {ticketCounts.finished}
-                  </div>
-                  <div className="text-xs text-gray-500">Finished</div>
-                </div>
+                <h2 className="text-xl font-semibold text-gray-900">Your Tickets</h2>
+                <span className="inline-flex items-center rounded-full bg-blue-100 px-2.5 py-0.5 text-xs font-medium text-blue-800">
+                  {filteredTickets.length} tickets
+                </span>
               </div>
             </div>
 
-            <div className="h-[560px] p-6">
-              {tickets.length === 0 ? (
+            <div className="p-6">
+              {filteredTickets.length === 0 ? (
                 <div className="py-12 text-center">
                   <div className="mx-auto mb-4 h-12 w-12 text-gray-400">
                     <svg
@@ -435,11 +269,15 @@ export const CtoTicketTest = () => {
                     </svg>
                   </div>
                   <h3 className="mb-2 text-lg font-medium text-gray-900">No tickets found</h3>
-                  <p className="text-gray-500">Create your first ticket to get started!</p>
+                  <p className="text-gray-500">
+                    {searchTerm
+                      ? 'No tickets match your search criteria.'
+                      : 'No tickets assigned to you yet.'}
+                  </p>
                 </div>
               ) : (
-                <div className="h-full space-y-4 overflow-y-auto">
-                  {tickets.map((ticket) => {
+                <div className="max-h-96 space-y-4 overflow-y-auto">
+                  {filteredTickets.map((ticket) => {
                     const isExpanded = expandedDescriptions.has(ticket.id);
                     const shouldTruncate = ticket.description.length > 120;
 
@@ -468,7 +306,7 @@ export const CtoTicketTest = () => {
                             </div>
                           </div>
                           <span className="ml-4 text-sm whitespace-nowrap text-gray-500">
-                            {formatDateTime(ticket.createdAt)}
+                            {getTimeAgo(ticket.createdAt)}
                           </span>
                         </div>
 
@@ -508,7 +346,7 @@ export const CtoTicketTest = () => {
                                 />
                               </svg>
                               <span>
-                                Assigned to: {ticket.employee.firstName} {ticket.employee.lastName}
+                                Assigned by: {ticket.author.firstName} {ticket.author.lastName}
                               </span>
                             </div>
                           </div>
@@ -523,14 +361,39 @@ export const CtoTicketTest = () => {
                                 strokeLinecap="round"
                                 strokeLinejoin="round"
                                 strokeWidth="2"
-                                d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"
+                                d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
                               />
                             </svg>
-                            <span>
-                              By: {ticket.author.firstName} {ticket.author.lastName}
-                            </span>
+                            <span>{formatDateTime(ticket.createdAt)}</span>
                           </div>
                         </div>
+
+                        {ticket.messages && ticket.messages.length > 0 && (
+                          <div className="mt-4 rounded-lg border border-gray-200 bg-gray-50 p-3">
+                            <div className="mb-1 flex items-center gap-2">
+                              <svg
+                                className="h-4 w-4 text-gray-500"
+                                fill="none"
+                                stroke="currentColor"
+                                viewBox="0 0 24 24"
+                              >
+                                <path
+                                  strokeLinecap="round"
+                                  strokeLinejoin="round"
+                                  strokeWidth="2"
+                                  d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z"
+                                />
+                              </svg>
+                              <span className="text-xs font-medium text-gray-700">
+                                Latest message:
+                              </span>
+                            </div>
+                            <p className="text-xs text-gray-600">
+                              {ticket.messages[0].content.substring(0, 80)}
+                              {ticket.messages[0].content.length > 80 ? '...' : ''}
+                            </p>
+                          </div>
+                        )}
                       </div>
                     );
                   })}
@@ -547,16 +410,17 @@ export const CtoTicketTest = () => {
                   <div className="flex items-center gap-3">
                     <div className="flex items-center gap-2">
                       <div className="h-3 w-3 animate-pulse rounded-full bg-green-500"></div>
-                      <span>Active Communication</span>
+                      <span>Active Chat</span>
                     </div>
                   </div>
                 ) : (
-                  'Select a ticket to communicate'
+                  'Select a ticket to chat'
                 )}
               </h2>
               {selectedTicket && (
                 <p className="mt-1 text-sm text-gray-600">
-                  with {selectedTicket.employee.firstName} {selectedTicket.employee.lastName}
+                  Communication with {selectedTicket.author.firstName}{' '}
+                  {selectedTicket.author.lastName}
                 </p>
               )}
             </div>
@@ -580,7 +444,7 @@ export const CtoTicketTest = () => {
                   </div>
 
                   <div className="mb-3 rounded-lg border border-gray-200 bg-white p-3">
-                    <h4 className="mb-1 text-sm font-medium text-gray-900">Task Description</h4>
+                    <h4 className="mb-1 text-sm font-medium text-gray-900">Description</h4>
                     <p className="max-h-16 overflow-y-auto text-xs leading-relaxed text-gray-700">
                       {selectedTicket.description}
                     </p>
@@ -602,8 +466,8 @@ export const CtoTicketTest = () => {
                         />
                       </svg>
                       <span>
-                        Assigned to: {selectedTicket.employee.firstName}{' '}
-                        {selectedTicket.employee.lastName}
+                        Assigned by: {selectedTicket.author.firstName}{' '}
+                        {selectedTicket.author.lastName}
                       </span>
                     </div>
                     <div className="flex items-center text-gray-500">
@@ -624,31 +488,15 @@ export const CtoTicketTest = () => {
                     </div>
                   </div>
 
-                  {selectedTicket.status === 'AWAITING_CONFIRMATION' && (
-                    <div className="mt-3 border-t border-gray-200 pt-3">
-                      <div className="flex items-center justify-between">
-                        <div className="flex items-center text-xs text-amber-600">
-                          <svg
-                            className="mr-1 h-3 w-3"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth="2"
-                              d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z"
-                            />
-                          </svg>
-                          <span>Employee marked this as finished</span>
-                        </div>
+                  {selectedTicket.status !== 'FINISHED' &&
+                    selectedTicket.status !== 'AWAITING_CONFIRMATION' && (
+                      <div className="mt-3 border-t border-gray-200 pt-3">
                         <button
                           onClick={handleMarkAsFinished}
-                          disabled={isConfirming}
+                          disabled={isMarkingAsFinished}
                           className="inline-flex items-center rounded-lg bg-green-600 px-3 py-2 text-xs font-medium text-white transition-colors hover:bg-green-700 focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50"
                         >
-                          {isConfirming ? (
+                          {isMarkingAsFinished ? (
                             <>
                               <svg
                                 className="mr-2 -ml-1 h-3 w-3 animate-spin text-white"
@@ -669,7 +517,7 @@ export const CtoTicketTest = () => {
                                   d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
                                 ></path>
                               </svg>
-                              Confirming...
+                              Marking as Finished...
                             </>
                           ) : (
                             <>
@@ -686,83 +534,10 @@ export const CtoTicketTest = () => {
                                   d="M5 13l4 4L19 7"
                                 />
                               </svg>
-                              Confirm Finished
+                              Mark as Finished
                             </>
                           )}
                         </button>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* CTO Direct Mark as Finished - Available for all non-finished tickets */}
-                  {selectedTicket.status !== 'FINISHED' &&
-                    selectedTicket.status !== 'AWAITING_CONFIRMATION' && (
-                      <div className="mt-3 border-t border-gray-200 pt-3">
-                        <div className="flex items-center justify-between">
-                          <div className="flex items-center text-xs text-blue-600">
-                            <svg
-                              className="mr-1 h-3 w-3"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth="2"
-                                d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-                              />
-                            </svg>
-                            <span>Mark as finished directly (CTO)</span>
-                          </div>
-                          <button
-                            onClick={handleMarkAsFinished}
-                            disabled={isMarkingFinished}
-                            className="inline-flex items-center rounded-lg bg-blue-600 px-3 py-2 text-xs font-medium text-white transition-colors hover:bg-blue-700 focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50"
-                          >
-                            {isMarkingFinished ? (
-                              <>
-                                <svg
-                                  className="mr-2 -ml-1 h-3 w-3 animate-spin text-white"
-                                  fill="none"
-                                  viewBox="0 0 24 24"
-                                >
-                                  <circle
-                                    className="opacity-25"
-                                    cx="12"
-                                    cy="12"
-                                    r="10"
-                                    stroke="currentColor"
-                                    strokeWidth="4"
-                                  ></circle>
-                                  <path
-                                    className="opacity-75"
-                                    fill="currentColor"
-                                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                                  ></path>
-                                </svg>
-                                Marking as Finished...
-                              </>
-                            ) : (
-                              <>
-                                <svg
-                                  className="mr-2 h-3 w-3"
-                                  fill="none"
-                                  stroke="currentColor"
-                                  viewBox="0 0 24 24"
-                                >
-                                  <path
-                                    strokeLinecap="round"
-                                    strokeLinejoin="round"
-                                    strokeWidth="2"
-                                    d="M5 13l4 4L19 7"
-                                  />
-                                </svg>
-                                Mark as Finished
-                              </>
-                            )}
-                          </button>
-                        </div>
                       </div>
                     )}
                 </div>
@@ -796,7 +571,7 @@ export const CtoTicketTest = () => {
                         </div>
                         <h3 className="mb-2 text-sm font-medium text-gray-900">No messages yet</h3>
                         <p className="text-xs text-gray-500">
-                          Start the conversation with your team member!
+                          Start the conversation with your manager!
                         </p>
                       </div>
                     </div>
@@ -815,7 +590,7 @@ export const CtoTicketTest = () => {
                               {/* Avatar */}
                               <div
                                 className={`flex h-6 w-6 items-center justify-center rounded-full text-xs font-medium text-white shadow-md ${
-                                  isCurrentUser ? 'bg-blue-500' : 'bg-gray-500'
+                                  isCurrentUser ? 'bg-green-500' : 'bg-blue-500'
                                 }`}
                               >
                                 {getInitials(
@@ -827,14 +602,14 @@ export const CtoTicketTest = () => {
                               <div
                                 className={`rounded-2xl px-3 py-2 shadow-sm ${
                                   isCurrentUser
-                                    ? 'rounded-br-md bg-blue-500 text-white'
+                                    ? 'rounded-br-md bg-green-500 text-white'
                                     : 'rounded-bl-md border bg-white text-gray-800'
                                 }`}
                               >
                                 <p className="text-sm leading-relaxed">{message.content}</p>
                                 <p
                                   className={`mt-1 text-xs ${
-                                    isCurrentUser ? 'text-blue-100' : 'text-gray-500'
+                                    isCurrentUser ? 'text-green-100' : 'text-gray-500'
                                   }`}
                                 >
                                   {formatTime(message.createdAt)}
@@ -900,7 +675,7 @@ export const CtoTicketTest = () => {
                         isSending ||
                         selectedTicket.status === 'FINISHED'
                       }
-                      className="inline-flex items-center rounded-lg bg-blue-600 px-4 py-2 font-medium text-white transition-colors hover:bg-blue-700 focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50"
+                      className="inline-flex items-center rounded-lg bg-green-600 px-4 py-2 font-medium text-white transition-colors hover:bg-green-700 focus:ring-2 focus:ring-green-500 focus:ring-offset-2 focus:outline-none disabled:cursor-not-allowed disabled:opacity-50"
                     >
                       {isSending ? (
                         <svg className="h-4 w-4 animate-spin" fill="none" viewBox="0 0 24 24">
@@ -956,10 +731,10 @@ export const CtoTicketTest = () => {
                     </svg>
                   </div>
                   <h3 className="mb-2 text-xl font-medium text-gray-900">
-                    Select a ticket to start communicating
+                    Select a ticket to start chatting
                   </h3>
                   <p className="text-gray-500">
-                    Choose a ticket from the list to chat with your team member
+                    Choose a ticket from the list to communicate with your manager
                   </p>
                 </div>
               </div>
