+++ app/routes/ProtectedRoute.tsx
@@ -2,6 +2,8 @@ import { Navigate } from 'react-router-dom';
 import routeNames from './route-names';
 import { Layout } from '@app/components/layout/Layout';
 import { useAuth } from '@app/context/AuthContext';
+import { defineAbilityFor } from '@app/casl/ability';
+import { AbilityContext } from '@app/casl/AbilityContext';
 
 export const ProtectedRoute = () => {
   const { isLoading, isAuthenticated } = useAuth();
