+++ app/routes/root.tsx
@@ -1,9 +1,7 @@
-import { QueryClientProvider } from '@tanstack/react-query';
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
-import { queryClient } from '../utils/query-client.ts';
 import routeNames from './route-names.tsx';
 import '../main.css';
-import AuthProvider from '@app/context/AuthContext.tsx';
+import { useAuth } from '@app/context/AuthContext.tsx';
 import { UnauthenticatedRoute } from './UnauthenticatedRoute.tsx';
 import { ProtectedRoute } from './ProtectedRoute.tsx';
 import { Layout } from '@app/components/layout/Layout.tsx';
@@ -27,8 +25,11 @@ import { UserDetail } from '@app/pages/UserDetail.tsx';
 import { Profile } from '@app/pages/Profile.tsx';
 import { Reports } from '@app/pages/Reports.tsx';
 import { ReportDetail } from '@app/pages/ReportDetail.tsx';
-import { CtoTicketTest } from '@app/pages/cto-ticket-test.tsx';
-import { EmployeeTicketTest } from '@app/pages/employee-ticket-test.tsx';
+import { TicketCTO } from '@app/pages/TicketCTO.tsx';
+import { TicketEmployee } from '@app/pages/TicketEmployee.tsx';
+import { NotAuthorized } from '@app/pages/NotAuthorized.tsx';
+import { defineAbilityFor } from '@app/casl/ability.ts';
+import { AbilityContext } from '@app/casl/AbilityContext.ts';
 
 const routesForPublic = [
   {
@@ -72,9 +73,9 @@ const routesForAuthenticated = [
       { path: routeNames.roles(), element: <Roles /> },
       { path: routeNames.reports(), element: <Reports /> },
       { path: routeNames.reportDetail(), element: <ReportDetail /> },
-      // Test pages - will be role-based later
-      { path: routeNames.ctoTicketTest(), element: <CtoTicketTest /> },
-      { path: routeNames.employeeTicketTest(), element: <EmployeeTicketTest /> },
+      { path: routeNames.ticketCTO(), element: <TicketCTO /> },
+      { path: routeNames.ticketEmployee(), element: <TicketEmployee /> },
+      { path: routeNames.notAuthorized(), element: <NotAuthorized /> },
     ],
   },
 ];
@@ -86,11 +87,12 @@ const router = createBrowserRouter([
 ]);
 
 export default function Root() {
+  const { user } = useAuth();
+  const ability = defineAbilityFor(user?.role?.permissions || []);
+
   return (
-    <QueryClientProvider client={queryClient}>
-      <AuthProvider>
-        <RouterProvider router={router} />
-      </AuthProvider>
-    </QueryClientProvider>
+    <AbilityContext.Provider value={ability}>
+      <RouterProvider router={router} />
+    </AbilityContext.Provider>
   );
 }
